Use pythran for native weight application
On large grids, application of the regridding weights can be very
expensive with the code as written. We can use pythran to compile this
function to native code and drastically speed it up. The way things
are currently written, it should fall back to the native code
implementation if the extension module isn't built.